==> metaG/R_analyses/Functions/AA_KO/08_DA_KO.Rmd <==
---
title: "KO DA"
author: "SRomano"
output: 
  html_document: 
    toc: yes
editor_options: 
  chunk_output_type: console
---
  
# Load packages

```{r setup}
list.of.packages <- c("tidyverse", "ggplot2", "RColorBrewer", "meta", "genodds")
new.packages <- list.of.packages[!(list.of.packages %in% installed.packages()[,"Package"])]
if(length(new.packages)) install.packages(new.packages)
if (!requireNamespace("BiocManager", quietly = TRUE)){
    install.packages("BiocManager")}


library(tidyverse)
library(genodds)
library(meta)


set.seed(9821)
```

# Load the data
This analysis needs to run on a named list in which each item is the abundance table of a study.

```{r}

l_df<-readRDS("/metaG/R_analyses/Functions/AA_KO/RDS/list_df_no0_KO.rds")
meta<-readRDS("/metaG/R_analyses/Taxonomy/RDS/meta.rds")
meta %>%
  group_by(Study, PD) %>%
  summarise(n = n()) %>%
  print( n = 50)

dim(meta)
lapply(l_df, dim)

#check the order is the same in all
x<-unlist(lapply(l_df, function(x) names(x)))
all(sort(x) == sort(meta$SampleID))
```

```{r}
# rename to use with the genodd function
names(meta)<-c("Sample_ID", "PD", "Other", "Cohort")
meta$PD<-factor(meta$PD, levels = c("HC", "PD"))
meta<-meta[order(meta$Sample_ID),]

# transpose for the DA
l_df<-lapply(l_df, function(x) as.data.frame(t(x)))

```

# Filter
```{r}
spec<-readRDS("/metaG/R_analyses/Functions/AA_KO/RDS/spec_5x10.rds")

l_df<-lapply(l_df, function(x) x[,which(names(x) %in% spec)])
l_df<-lapply(l_df, function(x) x[order(rownames(x)),])

```

# Caluclate DA
## GenOdds
```{r}
# source the function
source("/Scripts/DA_function_gen_odd.R")

df.cross<-calcualte_da_genOdds(list.df = l_df, 
                meta = meta,
                treat = "PD", assume_no_effect = TRUE)
saveRDS(df.cross, "/metaG/R_analyses/Functions/AA_KO/RDS/df.cross.KO.rds")

```

## Perform meta
```{r}
bins<-unique(df.cross$bin)

df<-data.frame(TE.random = vector(),
                 seTE.random = vector(),
                 lower.random = vector(),
                 upper.random = vector(),
                 pval.random = vector(),
                 bins = vector())
  
for(b in 1:length(bins)){
  require(meta)
    tmp.bin<-df.cross[df.cross$bin == bins[b],]
    print(b)
        
    print(dim(tmp.bin))

    x<-try(met<-metagen(TE=estimate, seTE=std.error,data=tmp.bin, method.tau = "REML",title =bins[b], studlab = Cohort))
    if(class(x)[1] != "try-error"){
    met.df<-as.data.frame(c(met[c("TE.random","seTE.random", "lower.random", "upper.random", "pval.random")],bins= bins[b])) 
    df<-rbind.data.frame(df, met.df)  
    } else { tmp<-data.frame(TE.random =NA, seTE.random = NA, lower.random = NA, upper.random = NA, pval.random = NA, bins= bins[b])
      df<-rbind.data.frame(df, tmp)}
}

df$p.adjust<-p.adjust(df$pval.random, method = "fdr")
```

## Subset sign
```{r}
subset(df,p.adjust<0.05 ) %>%
  dim()
df.sign<-subset(df,p.adjust<0.05 ) 

```

## Add description
```{r}
desc<-read.delim("KO.descr.def", quote = "", header = F, sep = "\t", )
# check length of KO
grep("KO", desc$V1) %>%
  length()

desc<-desc %>% 
  mutate(variable = rep(c("KO", "Def"), nrow(desc) / 2), 
         key = rep(1:(nrow(desc) / 2), each = 2)) %>%
  pivot_wider(id_cols = key, names_from = variable, values_from = V1)
dim(desc)
# There is one KO that has no entry in description
desc[grep("K22486", desc$KO),]
desc$KO<-gsub("ENTRY       K", "K", desc$KO)
desc$KO<-gsub("            .*KO", "", desc$KO)

desc$Def<-gsub("DEFINITION  ", "", desc$Def)
desc<-desc[,-1]

desc[grep("K22486", desc$KO),]


df.sign<-df.sign %>%
  left_join(desc, by = c("bins" = "KO"))


df.sign$dir<-ifelse(df.sign$TE.random<0, "CTRL", "PD")
saveRDS(df.sign, "/metaG/R_analyses/Functions/AA_KO/RDS/df.sign.KO.rds")
```

# Use gut modules
## Estimate modules
Need to reload the KO to use the unfiltered data set
```{r}
library(omixerRpm)
l_df<-readRDS("//metaG/R_analyses/Functions/AA_KO/RDS/list_df_no0_KO.rds")
# transpose it for the DA
l_df<-lapply(l_df, function(x) as.data.frame(t(x)))
lapply(l_df, rowSums)

# create another l with proper formatting for GM
l_df_gm<-l_df
for(i in 1:length(l_df)){
  bd<-as.data.frame(t(l_df[[i]]))
  l_df_gm[[i]]<-cbind.data.frame(entry = rownames(bd), bd)
}

# default DB
db <- loadDefaultDB()
r_df_gmm<-l_df
for(i in 1:length(l_df_gm)){
  mods <- rpm(l_df_gm[[i]], minimum.coverage=0.5, annotation = 1, module.db = db)
  r_df_gmm[[i]] <- asDataFrame(mods, "abundance")
}

# GBM
db <- loadDB("GBMs.v1.0")
r_df_gbm<-l_df
for(i in 1:length(l_df_gm)){
  mods <- rpm(l_df_gm[[i]], minimum.coverage=0.5, annotation = 1, module.db = db)
  r_df_gbm[[i]] <- asDataFrame(mods, "abundance")
}

```

## Calculae DA
### GBM
Sort GBM entries to avoid merging dfs with wrong variable order.

```{r}
db <- loadDB("GBMs.v1.0")
r_df_gbm<-lapply(r_df_gbm, function(x){
  tmp<-x[order(x$Module),]
  return(tmp)
})
gbm<-r_df_gbm
for(i in 1:length(r_df_gbm)){
  rownames(gbm[[i]])<-r_df_gbm[[i]]$Module
  gbm[[i]]<-as.data.frame(t(gbm[[i]][,c(-1,-2)]))
}
# transpose and save
gbm.t<-lapply(gbm, function(x) as.data.frame(t(x)))
gbm.t<-do.call(cbind.data.frame, gbm.t)

names(gbm.t)<-gsub("[A-Z][a-z]*_[0-9]*[.]", "", names(gbm.t)) 
names(gbm.t)<-gsub("Boktor_[1-2]_[0-9]*[.]", "", names(gbm.t)) 


gbm.t<-gbm.t[, order(names(gbm.t))]
meta<-meta[order(meta$Sample_ID),]
stopifnot(all(names(gbm.t) == meta$Sample_ID))
saveRDS(gbm.t, "/metaG/R_analyses/Functions/KO/RDS/GBM.rds")

df.gbm.go<-calcualte_da_genOdds(list.df = gbm, 
                meta = meta,
               treat = "PD", assume_no_effect = TRUE)
bins<-unique(df.gbm.go$bin)
saveRDS(df.gbm.go, "/metaG/R_analyses/Functions/AA_KO/RDS/df.cross.GBM.rds")

df.gbm<-data.frame(TE.random = vector(),
                 seTE.random = vector(),
                 lower.random = vector(),
                 upper.random = vector(),
                 pval.random = vector(),
                 bins = vector())
  
for(b in 1:length(bins)){
  require(meta)
    tmp.bin<-df.gbm.go[df.gbm.go$bin == bins[b],]
   x<-try(met<-metagen(TE=estimate, seTE=std.error,data=tmp.bin, method.tau = "REML",title =bins[b], studlab = Cohort))
    if(class(x)[1] != "try-error"){
    met.df<-as.data.frame(c(met[c("TE.random","seTE.random", "lower.random", "upper.random", "pval.random")],bins= bins[b])) 
    df.gbm<-rbind.data.frame(df.gbm, met.df)  
    } else { tmp<-data.frame(TE.random =NA, seTE.random = NA, lower.random = NA, upper.random = NA, pval.random = NA, bins= bins[b])
      df.gbm<-rbind.data.frame(df.gbm, tmp)}
}

# add description
nam<-db@module.names
df.gbm$descr<-vector(length = length(nrow(df.gbm)))
for(i in 1:nrow(df.gbm)){
  df.gbm$descr[i]<-nam[which(rownames(nam) == df.gbm$bins[i]), "V2"]
}

# adjust pvalue
df.gbm$p.adjust<-p.adjust(df.gbm$pval.random, method = "fdr")
df.sign<-subset(df.gbm, p.adjust < 0.05)
# save rds
saveRDS(df.sign, "/metaG/R_analyses/Functions/AA_KO/RDS/df.sign.GBM.rds")

```

### GMM
I need to reformat the df to remove the first 2 columns
```{r}
db <- loadDefaultDB()

gmm<-r_df_gmm
for(i in 1:length(r_df_gmm)){
  rownames(gmm[[i]])<-r_df_gmm[[i]]$Module
  gmm[[i]]<-as.data.frame(t(gmm[[i]][,c(-1,-2)]))
}
# transpose and save
gmm.t<-lapply(gmm, function(x) as.data.frame(t(x)))
gmm.t<-do.call(cbind.data.frame, gmm.t)

names(gmm.t)<-gsub("[A-Z][a-z]*_[0-9]*[.]", "", names(gmm.t)) 
names(gmm.t)<-gsub("Boktor_[1-2]_[0-9]*[.]", "", names(gmm.t)) 


gmm.t<-gmm.t[, order(names(gmm.t))]
meta<-meta[order(meta$Sample_ID),]
stopifnot(all(names(gmm.t) == meta$Sample_ID))
saveRDS(gmm.t, "/metaG/R_analyses/Functions/KO/RDS/GMM.rds")

 
df.gmm.go<-calcualte_da_genOdds(list.df = gmm, 
                meta = meta,
               treat = "PD", assume_no_effect = TRUE)
bins<-unique(df.gmm.go$bin)
saveRDS(df.gmm.go, "/metaG/R_analyses/Functions/AA_KO/RDS/df.cross.GMM.rds")

df.gmm<-data.frame(TE.random = vector(),
                 seTE.random = vector(),
                 lower.random = vector(),
                 upper.random = vector(),
                 pval.random = vector(),
                 bins = vector())

for(b in 1:length(bins)){
  require(meta)
    tmp.bin<-df.gmm.go[df.gmm.go$bin == bins[b],]
   x<-try(met<-metagen(TE=estimate, seTE=std.error,data=tmp.bin, method.tau = "REML",title =bins[b], studlab = Cohort))
    if(class(x)[1] != "try-error"){
    met.df<-as.data.frame(c(met[c("TE.random","seTE.random", "lower.random", "upper.random", "pval.random")],bins= bins[b])) 
    df.gmm<-rbind.data.frame(df.gmm, met.df)  
    } else { tmp<-data.frame(TE.random =NA, seTE.random = NA, lower.random = NA, upper.random = NA, pval.random = NA, bins= bins[b])
      df.gmm<-rbind.data.frame(df.gmm, tmp)}
}

# add description
nam<-db@module.names
df.gmm$descr<-vector(length = length(nrow(df.gmm)))
for(i in 1:nrow(df.gmm)){
  df.gmm$descr[i]<-nam[which(rownames(nam) == df.gmm$bins[i]), "V2"]
}

# adjust pvalue
df.gmm$p.adjust<-p.adjust(df.gmm$pval.random, method = "fdr")
df.sign<-subset(df.gmm, p.adjust < 0.05)
# save rds
saveRDS(df.sign, "/metaG/R_analyses/Functions/AA_KO/RDS/df.sign.GMM.rds")
```

# Save workspace
```{r}
d<-"/metaG/R_analyses/Functions/AA_KO/RData/"
save.image(paste0(d, "KO_DA_GenOdd.RData"))
```
